Initial PoC of the generator
This PoC shows that we can render UML diagrams in SVG format without
any user interaction.  Current output is stored in /output which will
contain both the sources as well as a rendered SVG.

It is our intention to manipulate the SVG so more information can be
added to it, ensuring information can be found in a visually
acceptable manner.