Preliminary support for inheritance tree.
This build supports the preliminary inheritance feature.  It is not
yet obvious whether we want to render all the properties or only the
explicitly defined properties on the given resource.  Current
rendering shows all properties.

Current build depends on a public feature build of mu-cl-resources as
inheritance is a non-released feature at this time.